Move the ValueRounding to the model
