+++ shared/data/model/src/test/java/com/ivy/data/model/util/ValueRoundingTest.kt
@@ -1,4 +1,4 @@
-package com.ivy.data.model.testing
+package com.ivy.data.model.util
 
 import com.google.testing.junit.testparameterinjector.TestParameter
 import com.google.testing.junit.testparameterinjector.TestParameterInjector
